Move type definitions to musicxmltypes.h/cpp Remove unused code from musicxml.h
