+++ src/importexport/musicxml/internal/musicxml/importmxmlpass1.cpp
@@ -43,6 +43,7 @@
 #include "importmxmllogger.h"
 #include "importmxmlnoteduration.h"
 #include "importmxmlpass1.h"
+#include "musicxmltypes.h"
 
 #include "modularity/ioc.h"
 #include "importexport/musicxml/imusicxmlconfiguration.h"
