+++ src/importexport/musicxml/internal/musicxml/importmxmlpass2.cpp
@@ -91,6 +91,7 @@
 #include "importmxmlpass2.h"
 #include "musicxmlfonthandler.h"
 #include "musicxmlsupport.h"
+#include "musicxmltypes.h"
 
 #include "modularity/ioc.h"
 #include "importexport/musicxml/imusicxmlconfiguration.h"
